requirements.txt
flask
flask_sqlalchemy
flask-cors
psycopg2-binary
python-dotenv
requests
gunicorn
openpyxl
